+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
--Features for next release
+
+### New Features
+
+- **AWSS3**
+   - Marks enumerate blocks methods as deprecated and provides and alternative which works with Swift. (See [PR #3726](https://github.com/aws-amplify/aws-sdk-ios/pull/3726))
 
 ## 2.25.0